Fixed a value in lossMap in ees1Descriptor in the test project.
esLossMapAssembler generates a loss map (array) and a corresponding array
with the maximum amount of time that the energy storage can charge or
discharge at any given power and state of charge. This will be used in the
dispatch and scheduling and checking operation.

ElectricalStorage updated to take max discharge time from the esLossMapAssembler.