Added time as an output of Demand class
limited src ask eess to provide to max it can in eesDispatch1.

Added a checkOperatingConditions in the init of ElectricalEnergyStorage, and changed the way it calculates under SRC

Added some out of bounds attributes to the ElectricalEnergyStorageSystem

Added genMolAvail attribute to Generator and Powerhouse

created predictLoad1, which tries to predict future load.

added simple unit scheduling in System Operations

added updates for wtgAvail in windFarm

called checkOperatingConditions in WindTurbine.