Added results for Set0 in Igiugig0
plotSet is in the process of being developed. It grabs results from the saved sql db for the set

getRunMetaData saves an sql db and a csv of the results for a set of simulations

generateRuns will not overwrite existing project data. It saves the actual values in setXComponentAttributes.db for attributes that are identical to another, instead of the other attribute name.