plotSetResult: added the ability to enter an equation for the plotted values
getRunMetaData: added the saving of the pdfs of overloading the diesel and GBS

generateRuns: added the ability to assign identical values from other components. For example, you can assign the same values of MOL to gen1 and gen2 as gen0. This prevents the generation of the different possible combinations if multiple MOL values are simulated. Previously it was only able to copy a value from within the same component (for example ees0 power in and power out max).