generateRuns: checks to see if value refers to a tag from the same component (instead of just component level) this is to make it backward compatable.
Generator: new function to updateGenPAvail. this is to avoid having to call updateOperatingConditions after running genSchedule.

Powerhouse: calculates the max start time of all generators. This will be used in the generator scheduling

SystemOperations: sets aside kWs energy when calculating the amount of SRC the eess can provide in the generator scheduling.