eesDispatch): and index of current step in order to checkOperatingConditions
ElectricalEnergyStorage: add tTindex to the input of eesDistpatch when calling

Powerhouse: add the simulation timeStep and subtract it from the required start or run time of each generator when scheduling. that is effectively counting the current time step as well as previous. this makes sense since the generator state is only updated for the beginning of the next time step.